+++ dotnet/src/Snowflake44.net461/Properties/AssemblyInfo.cs
@@ -3,9 +3,9 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("UID64")]
+[assembly: AssemblyTitle("Snowflake44")]
 [assembly: AssemblyDescription("a 44-Bit Timestamp with some random salt")]
-[assembly: AssemblyProduct("UID64")]
+[assembly: AssemblyProduct("Snowflake44")]
 
 [assembly: AssemblyCompany("(public domain)")]
 [assembly: AssemblyCopyright("SmartStandards")]
@@ -15,6 +15,6 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("46adc8b8-2f14-4a31-87ca-95b71ee09ea9")]
 
-[assembly: AssemblyVersion("1.0.0.*")]
-[assembly: AssemblyInformationalVersion("1.0.0-localbuild")]
+[assembly: AssemblyVersion("999.0.0")]
+[assembly: AssemblyInformationalVersion("999.0.0-localbuild")]
 //DONT ADD: [assembly: AssemblyFileVersion] !!!