+++ dotnet/src/Snowflake44/System/SmartStandards/Snowflake44.cs
@@ -25,6 +25,18 @@ namespace System.SmartStandards {
     /// </summary>
     /// <param name="incomingDate"> A date value between 1900-01-01 and 2457-01-01 </param>
     public static long EncodeDateTime(DateTime incomingDate) {
+      return EncodeDateTime(incomingDate, false);
+    }
+
+    /// <summary>
+    ///   Encodes a DateTime value to an integer 64 value.
+    /// </summary>
+    /// <param name="incomingDate"> A date value between 1900-01-01 and 2457-01-01 </param>
+    /// <param name="skipRandomSaltAndHistory">
+    ///   opt-out the addition of the random-salt portion, 
+    ///   which will result in the lowest possible long value for the given timestamp (for filtering purpose)
+    /// </param>
+    internal static long EncodeDateTime(DateTime incomingDate, bool skipRandomSaltAndHistory) {
 
       // 11
       // 98                 0
@@ -41,24 +53,35 @@ namespace System.SmartStandards {
           throw new Exception("Time stamp exceeds 44 bit!");
         }
 
-        if (elapsedMilliseconds != _PreviousTimeFrame) _RandomsOfCurrentTimeFrame.Clear();
+        lock (_RandomsOfCurrentTimeFrame) {
 
-        _PreviousTimeFrame = elapsedMilliseconds;
+          if (elapsedMilliseconds != _PreviousTimeFrame) {
+            _RandomsOfCurrentTimeFrame.Clear();
+          }
 
-        elapsedMilliseconds <<= 19;
+          _PreviousTimeFrame = elapsedMilliseconds;
 
-        int randomValue;
-        bool collisionDetected;
+          elapsedMilliseconds <<= 19;
 
-        do {
-          randomValue = RandomGenerator.Next(524287);
-          collisionDetected = !_RandomsOfCurrentTimeFrame.Add(randomValue);
-        } while (collisionDetected);
+          int randomValue = 0;
 
-        long randomPart = randomValue;
-        long uid = elapsedMilliseconds | randomPart;
+          if (!skipRandomSaltAndHistory) {
 
-        return uid;
+            bool collisionDetected;
+
+            do {
+              randomValue = RandomGenerator.Next(524287);
+              collisionDetected = !_RandomsOfCurrentTimeFrame.Add(randomValue);
+            } while (collisionDetected);
+
+          }
+
+          long randomPart = randomValue;
+          long uid = elapsedMilliseconds | randomPart;
+
+          return uid;
+        }
+
       }
 
     }
@@ -75,6 +98,39 @@ namespace System.SmartStandards {
       var decodedDate = new DateTime(CenturyBegin.Ticks + decodedTicks);
 
       return decodedDate;
+    }
+
+    /// <param name="from"></param>
+    /// <param name="before"></param>
+    /// <param name="fromUid"></param>
+    /// <param name="beforeUid">
+    ///  WARNING: this is time sensitive - consider to pass the date-portion only for correct mapping
+    ///  and this is representing the first value AFTER the selected range and will NOT be a match!
+    /// </param>
+    public static void GetTimeRange(
+      DateTime from, DateTime before,
+      out long fromUid, out long beforeUid
+    ) {
+
+      fromUid = Snowflake44.EncodeDateTime(from, skipRandomSaltAndHistory: true);
+      beforeUid = Snowflake44.EncodeDateTime(before, skipRandomSaltAndHistory: true);
+
+    }
+
+    /// <param name="extendee"></param>
+    /// <param name="from"></param>
+    /// <param name="before">
+    ///  WARNING: this is time sensitive - consider to pass the date-portion only for correct mapping
+    ///  and this is representing the first value AFTER the selected range and will NOT be a match!
+    /// </param>
+    public static bool IsInTimeRange(long extendee, DateTime from, DateTime before) {
+
+      Snowflake44.GetTimeRange(
+       from, before,
+       out long fromUid, out long untilUid
+     );
+
+      return (extendee >= fromUid && extendee < untilUid);
     }
 
   }